DCR-240 : handled order, other address, location preference of contact

+++ src/main/java/com/winsupply/mdmcustomertoecomsubscriber/repositories/OrderEmailAddressRepository.java
@@ -18,5 +18,5 @@ public interface OrderEmailAddressRepository extends JpaRepository<OrderEmailAdd
      *
      * @param pAddressId - the Address Id
      */
-    void deleteAllByAddressId(Long pAddressId);
+    void deleteAllByAddressId(Integer pAddressId);
 }
